requirements.txt
et-xmlfile==1.1.0
openpyxl==3.1.2
Pillow==9.5.0
pyglet==2.0.6
XlsxWriter==3.1.0
